Add NPU probe testing and transcription tools for Snapdragon
- Introduced `test-npu-probe.ps1` to validate the functionality of npu_probe.exe with various test cases, ensuring proper error handling and output verification.
- Added `transcribe-long-form.ps1` for processing audio files into transcriptions, including features for segmenting audio, managing overlapping windows, and logging results.
- Implemented Windows-specific threading support in `thread.c`, providing necessary functions for thread management and synchronization.

+++ src/tools/shuf.c
@@ -19,7 +19,7 @@ static void seed_rng(void) {
     }
 
     unsigned long long seed = (unsigned long long)platform_get_epoch_time();
-    unsigned long long addr = (unsigned long long)(unsigned long)(&seed);
+    unsigned long long addr = (unsigned long long)(size_t)&seed;
 
     shuf_rng_state = seed ^ (addr << 1U) ^ 0x9e3779b97f4a7c15ULL;
     if (shuf_rng_state == 0ULL) {
